fix: resolve MT4 compile errors found by CI
- Import KeyboardOptions from androidx.compose.foundation.text (not ui.text.input)
- InstrumentSpec.roundToLongOrZero now returns Long (kotlin.math.round returns Double)
- buildExecutionContext is suspend and no longer calls a suspend fun inside runCatching

+++ app/src/main/java/com/foxtrader/app/domain/usecase/risk/InstrumentSpec.kt
@@ -61,6 +61,6 @@ data class InstrumentSpec(
 
     private fun Double.roundToLongOrZero(): Long {
         if (!this.isFinite()) return 0L
-        return kotlin.math.round(this)
+        return kotlin.math.round(this).toLong()
     }
 }
